+++ src/backend/glutin_backend.rs
@@ -107,11 +107,6 @@ impl GlutinFacade {
         }
     }
 
-    /// Returns true if the window has been closed.
-    pub fn is_closed(&self) -> bool {
-        Option::as_ref(&self.backend).map(|b| b.borrow().is_closed()).unwrap_or(false)
-    }
-
     /// Returns the underlying window, or `None` if glium uses a headless context.
     pub fn get_window(&self) -> Option<WinRef> {
         Option::as_ref(&self.backend).map(|w| WinRef(w.borrow()))
@@ -257,10 +252,6 @@ impl GlutinWindowBackend {
 
     pub fn get_window(&self) -> &glutin::Window {
         &self.window
-    }
-
-    pub fn is_closed(&self) -> bool {
-        self.window.is_closed()
     }
 
     pub fn poll_events(&self) -> glutin::PollEventsIterator {